[#162595576] Change quantities to Double for OrderLeg and Fill

+++ src/main/java/it/trade/model/reponse/Fill.java
@@ -18,7 +18,7 @@ public class Fill {
 
     @SerializedName("quantity")
     @Expose
-    public Integer quantity;
+    public Double quantity;
 
     @Override
     public String toString() {
